Refactored Settings.
Created a new directory called 'conf' under the home package which
contains static files useful for setting up the Home project. These are
the bootstrap.sh provisioning bash script, supervisord.conf and a sample
config for Home.

Two new management commands have been added to output the contents of
the supervisord.conf and sample config. This allows the user to do:

    home conf_sample > myconf.py
    home supervisor_sample > /etc/supervisor/conf.d/home.conf

The change also refactors the logging setup, it is now initialised at
the main entry point of home rather than in the config. This allows for
more configurable logging while still allowing the user to replace the
full logging setup.