Expire values after 7 days.
For now, we will go with a seven day limit for data, once it gets to
that age we will access it via postgres. We may want to change this
later but its a decent start for now. The web UI only shows the last 24
hours at the moment, nevermind the last dy or the last 7.